requirements.txt
attrs==19.3.0
Automat==0.8.0
awscli==1.25.53
blinker==1.4
botocore==1.27.53
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==20.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
entrypoints==0.3
filelock==3.7.1
Flask==1.1.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==1.0.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
Pillow==7.0.0
platformdirs==2.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.1
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
PyYAML==5.4.1
requests==2.22.0
requests-unixsocket==0.2.0
roman==2.0.0
rsa==4.7.2
s3transfer==0.6.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
ssh-import-id==5.10
systemd-python==234
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.11
virtualenv==20.14.1
wadllib==1.3.3
Werkzeug==0.16.1
zipp==1.0.0
zope.interface==4.7.1
gunicorn==20.1.0
flask-sqlalchemy
python-dotenv
psycopg2-binary
python-jose
flask_wtf